Calculate application rates for active ingredients

Calculate application rates for active ingredients

Usage

application_rate_g_per_ha(
  product_uses,
  aggregation = c("max", "mean", "min"),
  dosage_units = c("percent_ww", "percent_vv", "state_of_matter"),
  skip_l_per_ha_without_g_per_L = TRUE,
  fix_l_per_ha = TRUE
)

Arguments

product_uses

A tibble containing the columns 'wNbr', 'use_nr', 'application_area_de', 'min_dosage', 'max_dosage', 'min_rate', 'max_rate', from the 'uses' table in a psmv_dm object, as well as the columns 'percent' and 'g_per_L' from the 'ingredients' table in a psmv_dm object.

aggregation

How to represent a range if present, e.g. "max" (default) or "mean".

dosage_units

If no units are given, or units are "%", then the applied amount in g/ha is calculated using a reference application volume and the dosage. As the dosage units are not explicitly given, we can specify our assumptions about these using this argument (currently not implemented, i.e. specifying the argument has no effect).

skip_l_per_ha_without_g_per_L

Per default, uses where the use rate has units of l/ha are skipped, if there is not product concentration in g/L. This was also done in the 2023 indicator project.

fix_l_per_ha

During the review of the 2023 indicator project calculations, a number of cases were identified where the unit l/ha specifies a water volume, and not a product volume. If TRUE (default), these cases are corrected, if FALSE, these cases are discarded.

Value

A tibble containing one additional column 'rate_g_per_ha'

Note

The reference application volume is used if there is no 'expenditure'. It is selected only based on the product application area. This is not correct if hops ('Hopfen') is the culture, as it has a unique reference application volume of 3000 L/ha. Applications to hops were excluded for calculating mean use rates in the indicator project (Korkaric 2023), arguing that it is not grown in large areas in Switzerland.

Examples

# \dontrun{
library(psmv)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(dm)
#> 
#> Attaching package: ‘dm’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
psmv_cur <- psmv_dm()

product_uses_with_ingredients <- psmv_cur |>
  dm_filter(substances =
    (substance_de %in% c("Halauxifen-methyl", "Kupfer (als Kalkpr\u00E4parat)"))) |>
  dm_flatten_to_tbl(uses, products) |>
  left_join(psmv_cur$ingredients, by = join_by(wNbr),
    relationship = "many-to-many") |>
  left_join(psmv_cur$substances, by = join_by(pk)) |>
  select(wNbr, name, use_nr,
    min_dosage, max_dosage, min_rate, max_rate, units_de,
    application_area_de,
    substance_de, percent, g_per_L)

application_rate_g_per_ha(product_uses_with_ingredients) |>
  filter(name %in% c("Cerelex", "Pixxaro EC", "Bordeaux S")) |>
  select(ai = substance_de, app_area = application_area_de,
    min_d = min_dosage,  max_d = max_dosage,
    min_r = min_rate, max_r = max_rate,
    units_de, rate = rate_g_per_ha) |>
  print(n = Inf)
#> # A tibble: 15 × 8
#>    ai                 app_area min_d max_d min_r max_r units_de   rate
#>    <chr>              <chr>    <dbl> <dbl> <dbl> <dbl> <chr>     <dbl>
#>  1 Cloquintocet-mexyl Feldbau     NA    NA  0.75    NA l/ha       4.5 
#>  2 Florasulam         Feldbau     NA    NA  0.75    NA l/ha       3.75
#>  3 Halauxifen-methyl  Feldbau     NA    NA  0.75    NA l/ha       4.72
#>  4 Cloquintocet-mexyl Feldbau     NA    NA  1       NA l/ha       6   
#>  5 Florasulam         Feldbau     NA    NA  1       NA l/ha       5   
#>  6 Halauxifen-methyl  Feldbau     NA    NA  1       NA l/ha       6.3 
#>  7 Cloquintocet-mexyl Feldbau     NA    NA  1       NA l/ha       6   
#>  8 Florasulam         Feldbau     NA    NA  1       NA l/ha       5   
#>  9 Halauxifen-methyl  Feldbau     NA    NA  1       NA l/ha       6.3 
#> 10 Fluroxypyr         Feldbau     NA    NA  0.5     NA l/ha     140   
#> 11 Cloquintocet-mexyl Feldbau     NA    NA  0.5     NA l/ha       6   
#> 12 Halauxifen-methyl  Feldbau     NA    NA  0.5     NA l/ha       6.25
#> 13 Fluroxypyr         Feldbau     NA    NA  0.5     NA l/ha     140   
#> 14 Cloquintocet-mexyl Feldbau     NA    NA  0.5     NA l/ha       6   
#> 15 Halauxifen-methyl  Feldbau     NA    NA  0.5     NA l/ha       6.25
# }