Calculate application rates for active ingredients

Calculate application rates for active ingredients

Usage

psmv_xml_application_rate_g_per_ha(
  product_uses,
  aggregation = c("mean", "max", "min")
)

Arguments

product_uses

A tibble containing the columns 'application_area_de', 'min_dosage', 'max_dosage', 'min_rate', 'max_rate', from the 'uses' table in a psmv_dm object, as well as the columns 'percent' and 'g_per_L' from the 'ingredients' table in a psmv_dm object.

aggregation

How to represent a range if present, e.g. "mean" or "max"

Value

A tibble containing one additional column 'rate_g_per_ha'

Note

The reference application volume is used if there is no 'expenditure'. It is selected only based on the product application area. This is not correct if hops ('Hopfen') is the culture, as it has a unique reference application volume of 3000 L/ha. Applications to hops were excluded for calculating mean use rates in the indicator project (Korkaric 2023), arguing that it is not grown in large areas in Switzerland.

Examples

library(psmv)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(dm)
#> 
#> Attaching package: ‘dm’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
psmv <- psmv_list[["2022"]]
  
product_uses_with_ingredients <- psmv |>
  dm_filter(substances = 
    (substance_de %in% c("Halauxifen-methyl", "Kupfer (als Kalkpr\u00E4parat)"))) |>
  dm_flatten_to_tbl(uses, products) |>
  left_join(psmv$ingredients, by = join_by(wNbr),
    relationship = "many-to-many") |>
  left_join(psmv$substances, by = join_by(pk)) |>
  select(wNbr, name, use_nr, 
    min_dosage, max_dosage, min_rate, max_rate, units_de,
    application_area_de, 
    substance_de, percent, g_per_L)

psmv_xml_application_rate_g_per_ha(product_uses_with_ingredients) |>
  filter(name %in% c("Cerelex", "Pixxaro EC", "Bordeaux S")) |>
  select(ai = substance_de, app_area = application_area_de,
    min_d = min_dosage,  max_d = max_dosage,
    min_r = min_rate, max_r = max_rate,
    units_de, rate = rate_g_per_ha) |>
  print(n = Inf)
#> # A tibble: 34 × 8
#>    ai                        app_area   min_d max_d min_r max_r units_de    rate
#>    <chr>                     <chr>      <dbl> <dbl> <dbl> <dbl> <chr>      <dbl>
#>  1 Kupfer (als Kalkpräparat) Gemüsebau  1.3    1.8  NA     NA   NA       3100   
#>  2 Kupfer (als Kalkpräparat) Gemüsebau  1.3    1.8  NA     NA   NA       3100   
#>  3 Kupfer (als Kalkpräparat) Weinbau    0.75  NA    12     NA   kg/ha    2400   
#>  4 Kupfer (als Kalkpräparat) Gemüsebau  0.5   NA     5     NA   kg/ha    1000   
#>  5 Kupfer (als Kalkpräparat) Weinbau    0.25  NA     4     NA   kg/ha     800   
#>  6 Kupfer (als Kalkpräparat) Gemüsebau NA     NA    13     NA   kg/ha    2600   
#>  7 Kupfer (als Kalkpräparat) Beerenbau  0.25   0.75  2.5    7.5 kg/ha    1000   
#>  8 Kupfer (als Kalkpräparat) Gemüsebau NA     NA     5     NA   kg/ha    1000   
#>  9 Kupfer (als Kalkpräparat) Beerenbau  0.5    1.3   5     13   kg/ha    1800   
#> 10 Kupfer (als Kalkpräparat) Gemüsebau NA     NA    13     18   kg/ha    3100   
#> 11 Kupfer (als Kalkpräparat) Obstbau    0.125  0.25  2      4   kg/ha     600   
#> 12 Kupfer (als Kalkpräparat) Gemüsebau NA     NA    13     18   kg/ha    3100   
#> 13 Kupfer (als Kalkpräparat) Obstbau    0.25   0.5   4      8   kg/ha    1200   
#> 14 Kupfer (als Kalkpräparat) Gemüsebau  0.5   NA     5     NA   kg/ha    1000   
#> 15 Kupfer (als Kalkpräparat) Obstbau    0.5    0.75  8     12   kg/ha    2000   
#> 16 Kupfer (als Kalkpräparat) Gemüsebau NA     NA    13     18   kg/ha    3100   
#> 17 Kupfer (als Kalkpräparat) Obstbau    0.5    0.75  8     12   kg/ha    2000   
#> 18 Kupfer (als Kalkpräparat) Beerenbau  0.25   0.75  2.5    7.5 kg/ha    1000   
#> 19 Kupfer (als Kalkpräparat) Feldbau   NA     NA    17.5   NA   kg/ha    3500   
#> 20 Fluroxypyr                Feldbau   NA     NA     0.5   NA   l/ha      140   
#> 21 Cloquintocet-mexyl        Feldbau   NA     NA     0.5   NA   l/ha        6   
#> 22 Halauxifen-methyl         Feldbau   NA     NA     0.5   NA   l/ha        6.25
#> 23 Fluroxypyr                Feldbau   NA     NA     0.5   NA   l/ha      140   
#> 24 Cloquintocet-mexyl        Feldbau   NA     NA     0.5   NA   l/ha        6   
#> 25 Halauxifen-methyl         Feldbau   NA     NA     0.5   NA   l/ha        6.25
#> 26 Cloquintocet-mexyl        Feldbau   NA     NA     0.75  NA   l/ha        4.5 
#> 27 Florasulam                Feldbau   NA     NA     0.75  NA   l/ha        3.75
#> 28 Halauxifen-methyl         Feldbau   NA     NA     0.75  NA   l/ha        4.72
#> 29 Cloquintocet-mexyl        Feldbau   NA     NA     1     NA   l/ha        6   
#> 30 Florasulam                Feldbau   NA     NA     1     NA   l/ha        5   
#> 31 Halauxifen-methyl         Feldbau   NA     NA     1     NA   l/ha        6.3 
#> 32 Cloquintocet-mexyl        Feldbau   NA     NA     1     NA   l/ha        6   
#> 33 Florasulam                Feldbau   NA     NA     1     NA   l/ha        5   
#> 34 Halauxifen-methyl         Feldbau   NA     NA     1     NA   l/ha        6.3