Get Parallel Imports from an XML version of the PSMV

Get Parallel Imports from an XML version of the PSMV

Usage

psmv_xml_get_parallel_imports(psmv_xml = psmv_xml_get())

Arguments

psmv_xml

An object as returned by 'psmv_xml_get'

Value

A tibble with a row for each parallel import section in the XML file.

Examples

library(psmv)
parallel_imports_2015 <- psmv_xml_get(2015) |>
  psmv_xml_get_parallel_imports()
#> ℹ Reading XML for 2015-01-06
#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if
#> `.name_repair` is omitted as of tibble 2.0.0.
#> ℹ Using compatibility `.name_repair`.
#> ℹ The deprecated feature was likely used in the psmv package.
#>   Please report the issue at <https://gitlab.agsad.admin.ch/psmwb/psmv/issues>.

# Get current list of parallel_imports
psmv_xml_get_parallel_imports()
#> ℹ Reading XML for 2024-02-12
#> # A tibble: 749 × 9
#>    id     wNbr  name   admissionnumber producingCountryPrim…¹ exhaustionDeadline
#>    <chr>  <chr> <chr>  <chr>           <chr>                  <chr>             
#>  1 F-7056 1899  Life … 2159999         6171                   ""                
#>  2 I-2789 1899  Pirim… 4701            6176                   ""                
#>  3 B-5440 2054  Basam… 5675P/B         6163                   ""                
#>  4 F-3632 2054  Basam… 6800403         6171                   ""                
#>  5 I-6151 2054  Basam… 1573            6176                   ""                
#>  6 F-2010 2671  Fluid… 5100219         6171                   ""                
#>  7 F-7271 2935  Polyr… 8700107         6171                   ""                
#>  8 I-7270 2935  Polyr… 7916            6176                   ""                
#>  9 I-5565 3002  Dazid… 012455          6176                   ""                
#> 10 I-6469 3049  Bonzi  016305          6176                   ""                
#> # ℹ 739 more rows
#> # ℹ abbreviated name: ¹​producingCountryPrimaryKey
#> # ℹ 3 more variables: soldoutDeadline <chr>, packageInsert <chr>,
#> #   permission_holder_key <chr>