Remove product sections with duplicated W-Numbers

Remove product sections with duplicated W-Numbers

Usage

psmv_xml_remove_duplicated_wNbrs(
  psmv_xml = psmv_xml_get(),
  keep = c("last", "first")
)

Arguments

psmv_xml

An object as returned by psmv_xml_get

keep

How should we deal with product sections with identical W-Numbers? The default strategy 'keep' only keeps the last of such sets of product sections.

Value

An 'psmv_xml' object with only product sections that have unique W-Numbers

Examples

psmv_xml_remove_duplicated_wNbrs()
#> ℹ Reading XML for 2023-11-01
#> {xml_document}
#> <PublicationData>
#>  [1] <Products numberOfProducts="1677">\n  <Product id="7638" wNbr="6509" nam ...
#>  [2] <Parallelimports numberOfParallelimports="700">\n  <Parallelimport id="D ...
#>  [3] <MetaData name="ApplicationArea" numberOfRows="10">\n  <Detail primaryKe ...
#>  [4] <MetaData name="ApplicationComment" numberOfRows="740">\n  <Detail prima ...
#>  [5] <MetaData name="CultureForm" numberOfRows="2">\n  <Detail primaryKey="64 ...
#>  [6] <MetaData name="Culture" numberOfRows="315">\n  <Detail primaryKey="1139 ...
#>  [7] <MetaData name="FormulationCode" numberOfRows="44">\n  <Detail primaryKe ...
#>  [8] <MetaData name="Measure" numberOfRows="96">\n  <Detail primaryKey="6571" ...
#>  [9] <MetaData name="Obligation" numberOfRows="1398">\n  <Detail primaryKey=" ...
#> [10] <MetaData name="Pest" numberOfRows="522">\n  <Detail primaryKey="11956"> ...
#> [11] <MetaData name="ProductCategory" numberOfRows="26">\n  <Detail primaryKe ...
#> [12] <MetaData name="TimeMeasure" numberOfRows="2">\n  <Detail primaryKey="73 ...
#> [13] <MetaData name="Substance" numberOfRows="433">\n  <Detail primaryKey="3" ...
#> [14] <MetaData name="PermissionHolder" numberOfRows="214">\n  <PermissionHold ...
#> [15] <MetaData name="City" numberOfRows="181">\n  <Detail primaryKey="3285">\ ...
#> [16] <MetaData name="Country" numberOfRows="20">\n  <Detail primaryKey="6161" ...
#> [17] <MetaData name="CodeR" numberOfRows="55">\n  <Detail primaryKey="12045"> ...
#> [18] <MetaData name="CodeS" numberOfRows="31">\n  <Detail primaryKey="6039">\ ...
#> [19] <MetaData name="DangerSymbol" numberOfRows="8">\n  <Detail primaryKey="8 ...
#> [20] <MetaData name="SignalWords" numberOfRows="2">\n  <Detail primaryKey="79 ...
#> ...