Create a dm object from an XML version of the PSMV

Create a dm object from an XML version of the PSMV

Usage

psmv_dm(from = psmv_xml_url, remove_duplicates = TRUE)

# S3 method for psmv_dm
print(x, ...)

Arguments

from

A specification of the way to retrieve the XML

remove_duplicates

Should duplicates based on wNbrs be removed?

x

A psmv_dm object

...

Not used

Value

A dm object with tables linked by foreign keys pointing to primary keys, i.e. with referential integrity.

Examples

# \dontrun{
library(dm)

psmv_cur <- psmv_dm()
dm_examine_constraints(psmv_cur)
#> Adding missing grouping variables: `pNbr`
#> ! Unsatisfied constraints:
#> • Table `CodeR`: foreign key `wNbr` into table `products`: values of `CodeR$wNbr` not in `products$wNbr`: 6112 (10), 6403 (10), 6403-1 (10), 6403-2 (10), 6403-3 (10), …
#> • Table `CodeS`: foreign key `wNbr` into table `products`: values of `CodeS$wNbr` not in `products$wNbr`: 4574 (7), 4574-1 (7), 5745 (7), 5745-2 (7), 5745-3 (7), …
#> • Table `danger_symbols`: foreign key `wNbr` into table `products`: values of `danger_symbols$wNbr` not in `products$wNbr`: 4491 (5), 4491-1 (5), 5365 (5), 6090 (5), 6090-1 (5), …
#> • Table `formulation_codes`: foreign key `wNbr` into table `products`: values of `formulation_codes$wNbr` not in `products$wNbr`: 4904 (2), 7377 (2), 1454 (1), 1454-2 (1), 1526 (1), …
#> • Table `ingredients`: foreign key `wNbr` into table `products`: values of `ingredients$wNbr` not in `products$wNbr`: 6362 (8), 4687 (7), 4687-1 (7), 4687-2 (7), 4687-3 (7), …
#> • Table `parallel_imports`: foreign key `wNbr` into table `products`: values of `parallel_imports$wNbr` not in `products$wNbr`: 5481 (26), 5218 (16), 4495 (14), 6747 (13), 4701 (12), …
#> • Table `product_categories`: foreign key `wNbr` into table `products`: values of `product_categories$wNbr` not in `products$wNbr`: 2054 (4), 4555 (3), 4678 (3), 5168 (3), 5168-2 (3), …
#> • Table `signal_words`: foreign key `wNbr` into table `products`: values of `signal_words$wNbr` not in `products$wNbr`: 1454 (1), 1454-2 (1), 1526 (1), 1526-1 (1), 1526-2 (1), …
#> • Table `uses`: foreign key `wNbr` into table `products`: values of `uses$wNbr` not in `products$wNbr`: 6880 (85), 6880-2 (85), 6880-3 (85), 6020 (81), 6020-1 (81), …
dm_draw(psmv_cur)

# Show some information for products named 'Boxer'
psmv_cur |>
  dm_filter(products = (name == "Boxer")) |>
  dm_nrow()
#>             products   product_categories    formulation_codes 
#>                    0                    0                    0 
#>     parallel_imports       danger_symbols                CodeS 
#>                    0                    0                    0 
#>                CodeR         signal_words           substances 
#>                    0                    0                    0 
#>          ingredients                 uses application_comments 
#>                    0                    0                    0 
#>        culture_forms             cultures                pests 
#>                    0                    0                    0 
#>          obligations 
#>                    0 
# }