Create a dm object from an XML version of the PSMV

Create a dm object from an XML version of the PSMV

Usage

psmv_dm(
  from = last(psmv::psmv_xml_dates),
  remove_duplicates = TRUE,
  keep = c("last", "first")
)

# S3 method for psmv_dm
print(x, ...)

Arguments

from

A number giving a year starting from 2011 up to the current year, or one of the dates in psmv_xml_dates as a length one character vector in the format YYYY-MM-DD, or an URL from where to download

remove_duplicates

Should duplicates based on wNbrs be removed?

keep

How should we deal with product sections with identical W-Numbers? The default strategy 'keep' only keeps the last of such sets of product sections.

x

A psmv_dm object

...

Not used

Value

A dm object with tables linked by foreign keys pointing to primary keys, i.e. with referential integrity.

Examples

library(dm)
#> 
#> Attaching package: ‘dm’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
psmv_2017 <- psmv_dm(2017)
#> ℹ Reading XML for 2017-01-13
#> ✔ No duplicated W-Numbers
print(psmv_2017)
#> <psmv_dm> object from: 2017-01-13 
#>             products   product_categories    formulation_codes 
#>                 2269                 2488                 2269 
#>       danger_symbols                CodeS                CodeR 
#>                 3400                10051                 6659 
#>         signal_words           substances          ingredients 
#>                  976                  534                 3618 
#>                 uses application_comments        culture_forms 
#>                15642                 7994                  861 
#>             cultures                pests          obligations 
#>                24277                28729                45255 
dm_examine_constraints(psmv_2017)
#> Adding missing grouping variables: `pNbr`
#> ℹ All constraints satisfied.
# Show some information for products named 'Boxer'
psmv_2017 |>
  dm_filter(products = (name == "Boxer")) |>
  dm_nrow()
#>             products   product_categories    formulation_codes 
#>                    2                    2                    2 
#>       danger_symbols                CodeS                CodeR 
#>                    4                    4                    8 
#>         signal_words           substances          ingredients 
#>                    2                    1                    2 
#>                 uses application_comments        culture_forms 
#>                   14                   16                    2 
#>             cultures                pests          obligations 
#>                   24                   24                   70 
dm_draw(psmv_2017)

psmv_dm_cur <- psmv_dm(psmv_xml_url)
#> ✔ No duplicated W-Numbers